endpoints specified in secrets config not env var
